[CCUBE-1512-1][JY] Create common functions and helpers for generate schema logic
