+++ src/translator/index.ts
@@ -1,4 +1,3 @@
-export * from "./translator";
 export * from "./helper";
+export * from "./translator";
 export * from "./types";
-export * from "./text-based-field";
