+++ src/translator/text-based-field.ts
@@ -11,14 +11,9 @@ import {
     EValidationTypeFEE,
     IValidation,
     TElement,
-    TTextBasedElement,
 } from "src/context-providers";
 import { SimpleIdGenerator } from "src/util/simple-id-generator";
-import {
-    createConditionalRenderingObject,
-    parseConditionalRenderingObject,
-    parsePrefillObject,
-} from "./helper";
+import { parseConditionalRenderingObject, parsePrefillObject } from "./helper";
 import { IPrefillConfig } from "./types";
 
 export namespace TextBasedField {
@@ -33,26 +28,6 @@ export namespace TextBasedField {
     }
 
     namespace Email {
-        export const createEmailValidationSchema = (
-            validation: IValidation
-        ) => {
-            const domainRegexString = (domains: IValidation) => {
-                if (domains) {
-                    const domainsArr = domains?.validationRule.split(",");
-                    const generateSchemaDomains = domainsArr?.map((domain) =>
-                        domain.trim().replace(/^@/, "").replace(/\./g, "\\.")
-                    );
-                    const regexPattern = `^[a-zA-Z0-9._%+-]+@(${generateSchemaDomains.join("|")})$`;
-                    return new RegExp(regexPattern);
-                }
-            };
-
-            return {
-                matches: domainRegexString(validation).toString(),
-                errorMessage: validation.validationErrorMessage,
-            };
-        };
-
         export const parseEmailValidation = (
             validation: IYupValidationRule[]
         ) => {
@@ -79,46 +54,6 @@ export namespace TextBasedField {
             }
         };
     }
-
-    const createTextFieldValidationSchema = (validation: IValidation[]) => {
-        const validationObj = validation.reduce<ISchemaValidation[]>(
-            (acc, value) => {
-                switch (value.validationType) {
-                    case EValidationType.MAX_VALUE:
-                    case EValidationType.MAX_LENGTH:
-                        acc.push({
-                            max: parseInt(value.validationRule),
-                            errorMessage: value.validationErrorMessage,
-                        });
-                        break;
-                    case EValidationType.MIN_VALUE:
-                    case EValidationType.MIN_LENGTH:
-                        acc.push({
-                            min: parseInt(value.validationRule),
-                            errorMessage: value.validationErrorMessage,
-                        });
-                        break;
-                    case EValidationType.WHOLE_NUMBERS:
-                        acc.push({
-                            integer: true,
-                            errorMessage: value.validationErrorMessage,
-                        });
-                        break;
-                    case EValidationType.CUSTOM_REGEX:
-                        acc.push({
-                            matches: value.validationRule,
-                            errorMessage: value.validationErrorMessage,
-                        });
-                        break;
-                    default:
-                        break;
-                }
-                return acc;
-            },
-            []
-        );
-        return validationObj;
-    };
 
     export const parseTextBasedFieldValidation = (
         validation: IYupValidationRule[]
@@ -162,48 +97,6 @@ export namespace TextBasedField {
         return validationObj;
     };
 
-    const createValidationObject = (element: TTextBasedElement) => {
-        const validation: IYupValidationRule[] = [];
-
-        if (element.required) {
-            validation.push({
-                required: element.required,
-                ...(element.requiredErrorMsg && {
-                    errorMessage: element.requiredErrorMsg,
-                }),
-            });
-        }
-
-        if (element.validation && element.validation.length > 0) {
-            switch (element.type) {
-                case EElementType.EMAIL: {
-                    const validationChild = Email.createEmailValidationSchema(
-                        element.validation[0]
-                    );
-                    validation.push(validationChild);
-                    break;
-                }
-                case EElementType.TEXT:
-                case EElementType.TEXTAREA:
-                case EElementType.CONTACT:
-                case EElementType.NUMERIC: {
-                    const validationChild = createTextFieldValidationSchema(
-                        element.validation
-                    );
-                    validation.push(...validationChild);
-                    break;
-                }
-                case EElementType.CHECKBOX:
-                case EElementType.RADIO:
-                    break;
-            }
-        }
-
-        if (validation.length === 0) return;
-
-        return validation;
-    };
-
     export const parseValidationObject = (
         type: EElementType,
         validation: IYupValidationRule[]
@@ -221,39 +114,6 @@ export namespace TextBasedField {
                 return;
         }
     };
-
-    export const elementToSchema = (element: TTextBasedElement) => {
-        const conditionalRenderingObject = createConditionalRenderingObject(
-            element?.conditionalRendering
-        );
-        const validationObject = createValidationObject(element);
-        const textBasedFieldSchema = {
-            [element.id]: {
-                label: {
-                    mainLabel: element.label,
-                    ...(element.description && {
-                        subLabel: element.description,
-                    }),
-                },
-                uiType: element.type,
-                columns: {
-                    desktop: element.columns.desktop,
-                    tablet: element.columns.tablet,
-                    mobile: element.columns.mobile,
-                },
-                ...(element.placeholder && {
-                    placeholder: element.placeholder,
-                }),
-                ...(validationObject && { validation: validationObject }),
-                ...(conditionalRenderingObject && {
-                    showIf: conditionalRenderingObject,
-                }),
-            },
-        };
-
-        return textBasedFieldSchema;
-    };
-
     export const parseToElement = (
         element: TElementSchema,
         key: string,
